requirements.txt
Django>=5.0,<6.0
djangorestframework
djangorestframework-simplejwt
gunicorn
psycopg[binary]
dj-database-url
python-decouple